[SES-45] Moved accessor methods to a different file
