+++ src/stories/pages/page/cu-about/cu-about.stories.tsx
@@ -5,7 +5,7 @@ import { CardInfoMemberType } from '../../../components/card-info-member/card-in
 import { LinkModel } from '../../../components/cutable-column-links/cutable-column-links';
 import { RelateMipType } from '../../../components/relate-mips/relate-mips';
 import { CuStatusEnum } from '../../../../core/enums/cu-status.enum';
-import {LinkTypeEnum} from "../../../../core/enums/link-type.enum";
+import { LinkTypeEnum } from '../../../../core/enums/link-type.enum';
 
 export default {
   title: 'Pages/CuAbout',
